Use SHARED_CRYPTO_PATH for the shared crypto path
In CryptoBinPkg, some files (like Crypto.inc.dsc) are used during
the build. This should come from the latest revisions in the
mu_crypto_release repo (which doesn't use the actual binaries).

In consuming repos, the path will likely come from MU_BASECORE.
Tehcnically, the ext dep could be defined anywhere. To satisfy
both of these needs and make the location more portable in
general, an environment variable is introduced to track the
path.

In mu_crypto_release, the path will be set to "CryptoBinPkg". In
shared crypto consuming repos, it will be set in the ext dep to
the path the binaries are placed in.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>