Update nuget package bundling (#86)
## Description

Updates the logic that is responsible to bundling the nuget contents in
the appropriate manner. This update replaces the standalone script with
a post build plugin that will be executed when the command line argument
`--bundle` is added to either the MultiFlavorBuild or SingleFlavorBuild
build scripts.

Additionally, update the dsc to generate pdbs on release builds in
addition to debug builds. Also ensures pdb, map, and build available and
a part of the bundle that is generated for a NuGet release. These
changes make it easier for the local story, such that developers can run
the `py MultiFlavorBuild.py ... --bundle`, and the structure of the
NuGet package is generated. Instead of copying over the newly compiled
binaries to the NuGet package in the platform, the developer can simply
set the `SHARED_CRYPTO_PATH=<workspace/Bundle>` and build.

- [ ] Impacts functionality?
- **Functionality** - Does the change ultimately impact how firmware
functions?
- Examples: Add a new library, publish a new PPI, update an algorithm,
...
- [ ] Impacts security?
- **Security** - Does the change have a direct security impact on an
application,
    flow, or firmware?
  - Examples: Crypto algorithm change, buffer overflow fix, parameter
    validation improvement, ...
- [ ] Breaking change?
- **Breaking change** - Will anyone consuming this change experience a
break
    in build or boot behavior?
- Examples: Add a new library class, move a module to a different repo,
call
    a function in a new library class in a pre-existing module, ...
- [ ] Includes tests?
  - **Tests** - Does the change include any explicit test code?
  - Examples: Unit tests, integration tests, robot tests, ...
- [ ] Includes documentation?
- **Documentation** - Does the change contain explicit documentation
additions
    outside direct code modifications (and comments)?
- Examples: Update readme file, add feature readme file, link to
documentation
    on an a separate Web page, ...

## How This Was Tested

Ensured compiling, bundling, and releasing continues to work:
https://dev.azure.com/projectmu/mu/_build/results?buildId=69830&view=results

## Integration Instructions

N/A