[REVERT] Invalid File Exclusion (#93)
# Preface

## Description

Expectation is that the driver includes all INF files and that
MU_BASECORE removes them if necessary to prevent duplicates. This change
was made in error. The driver pack should not be excluding file types.


For each item, place an "x" in between `[` and `]` if true. Example:
`[x]`.
_(you can also check items in the GitHub UI)_

- [X] Impacts functionality?
  - Driver will now include the correct files (*.inf, *.uni)
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?
- [ ] 
## How This Was Tested

N/A

## Integration Instructions

N/A