Add Non-MM Supervisor Standalone MM X64 Binary
Builds a Standalone MM binary that can be used with StandaloneMmPkg
as opposed to the MmSupervisorPkg.

To provide forward looking clarity, the MM Supervisor instance is
renamed to include "MM Supervisor" when referenced. It is configured
in platform DSCs using the following macros:

- `STANDALONEMM_MMSUPV_CRYPTO_SERVICES`
- `STANDALONEMM_MMSUPV_CRYPTO_ARCH`

That makes this a breaking change. All platforms using the MM
Supervisor Standalone MM binaries must update their platforms
to use the MM Supervisor instance of the binary from this change
forward.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>