Update instructions for semantic versioning
Shared Crypto now follows [semantic versioning](https://semver.org/).
The version number is incremented based on the type of changes made
to the shared crypto binaries. The version number is not tied to the
version of the underlying crypto provider.

Prior to adopting semantic versinoning in the `1.0.0` release, the
version number followed a form of `YYYY.MM.PATCH`. The `1.0.0`
release was the first to use semantic versioning and proceeds any of
those versions.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>