Update instructions for semantic versioning (#115)
## Description

Closes #116 

Shared Crypto now follows [semantic versioning](https://semver.org/).
The version number is incremented based on the type of changes made to
the shared crypto binaries. The version number is not tied to the
version of the underlying crypto provider.

Prior to adopting semantic versioning in the `1.0.0` release, the
version number followed a form of `YYYY.MM.PATCH`. The `1.0.0` release
was the first to use semantic versioning and proceeds any of those
versions.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [x] Includes documentation?

## How This Was Tested

- CI

## Integration Instructions

- Following the release with this change semantic versioning will be
used. This change will be present in version `1.0.0`.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>