Drop stuart git dependencies
This repo currently has a mix of git dependencies and submodules. The
dependencies are pointing to release/202405 while the submodules are
referring to release/202502. This can cause a user to get the
incorrect MU_BASECORE content depending on how their workspace is
initialized.

Submodules seem to be the main path used, so remove the stuart
dependencies and add the submodules to the required submodule list.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>